backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.0
pandas==2.2.0
python-multipart==0.0.7
requests==2.31.0
pydantic-settings==2.1.0
langchain-google-genai==0.0.6
google-generativeai==0.3.2
